feat: alteracoes para o relacionamento livro fornecedor

+++ magiadoslivros/src/main/java/com/eternos/magiadoslivros/domain/repository/LivroRepository.java
@@ -4,6 +4,6 @@ import com.eternos.magiadoslivros.domain.model.Livro;
 
 import org.springframework.data.jpa.repository.JpaRepository;
     
-public interface LivroRepository extends JpaRepository<Livro, Long> {
+public interface LivroRepository extends JpaRepository<Livro, Integer> {
     
 }
